Refer to the Flatcar docs version by the major version
Releases that have the same major version are built from the same
maintenance branch where no big changes are made. Documentation
improvements don't have to be done for a particular release and it
could quickly be too much to handle. Just use the major version to
refer to the docs branch that was created when a new major version
was branched (but for now it's still "main" until we fix some basic
issues).